Rename the `light-client` to `web-client`
Rename the `light-client` crate to `web-client` and re-write the
client code to be `wasm` friendly.
