--- light-client/nimiq-client.service
@@ -1,14 +1,0 @@
-[Unit]
-Description=Nimiq's Rust light client
-After=syslog.target network.target remote-fs.target nss-lookup.target
-
-[Service]
-Type=exec
-ExecStartPre=$(which nimiq-light-client)
-ExecStart=$(which nimiq-light-client)
-ExecReload=/bin/kill -s HUP $MAINPID
-ExecStop=/bin/kill -s QUIT $MAINPID
-PrivateTmp=true
-
-[Install]
-WantedBy=multi-user.target
