+++ web-client/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "nimiq-light-client"
+name = "nimiq-web-client"
 version = "0.1.0"
 authors = ["The Nimiq Core Development Team <info@nimiq.com>"]
 edition = "2021"
-description = "Nimiq's Rust light client"
+description = "Nimiq's Rust to wasm web client"
 homepage = "https://nimiq.com"
 repository = "https://github.com/nimiq/core-rs-albatross"
 license = "Apache-2.0"
@@ -19,9 +19,10 @@ maintenance = { status = "experimental" }
 
 [dependencies]
 futures = { package = "futures-util", version = "0.3" }
+gloo-timers = "0.2"
 log = { package = "tracing", version = "0.1", features = ["log"] }
-tokio = { version = "1.24", features = ["macros", "rt-multi-thread", "time"] }
-tokio-metrics = "0.1"
+wasm-bindgen = "0.2"
+wasm-bindgen-futures = "0.4"
 
 [dependencies.nimiq]
 package = "nimiq-lib"
@@ -29,11 +30,6 @@ path = "../lib"
 version = "0.1"
 default-features = false
 features = [
-    "deadlock",
-    "logging",
-    "metrics-server",
     "panic",
-    "rpc-server",
-    "signal-handling",
-    "wallet",
+    "web-logging",
 ]
