+++ web-client/index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en-US">
+  <head>
+    <meta charset="utf-8" />
+    <title>Nimiq Web client example</title>
+  </head>
+  <body>
+    <script type="module">
+      import init, { nimiq_web_client } from "./pkg/nimiq_web_client.js";
+      init().then(() => {
+	    nimiq_web_client();
+      });
+    </script>
+  </body>
+</html>
+
